+++ marashov.alexander/A1/rectangle.hpp
@@ -6,14 +6,14 @@
 class Rectangle : public Shape
 {
   public:
-    Rectangle(const point_t& pos, float width, float height);
-    Rectangle(float posX, float posY, float width, float height);
+    Rectangle(const point_t& pos, double width, double height);
+    Rectangle(double posX, double posY, double width, double height);
 
-    float getArea() const override;
+    double getArea() const override;
     rectangle_t getFrameRect() const override;
     point_t getCenter() const override;
     void move(const point_t& pos) override;
-    void move(float dltX, float dltY) override;
+    void move(double dltX, double dltY) override;
 
   private:
     rectangle_t m_figure;
